microchange

+++ backend/internal/domain/log.go
@@ -7,14 +7,14 @@ type ExerciseLog struct {
 	Sets        []SetLog
 	Difficulty  int
 	Cycle       int
-	RPE         int `json:"rpe"` // Rate of Perceived Exertion (1-10)
+	RPE         int `json:"rpe"`         // Rate of Perceived Exertion (1-10)
 	FormQuality int `json:"formQuality"` // 1-5 scale
 }
 
 type SetLog struct {
-	Reps   int `json:"reps"`
-	Weight int `json:"weight"`
-	RPE    int `json:"rpe"` // RPE for this specific set (1-10)
+	Reps   int     `json:"reps"`
+	Weight float64 `json:"weight"`
+	RPE    float64 `json:"rpe"` // RPE for this specific set (1-10)
 }
 
 type WorkoutLog struct {
@@ -23,9 +23,9 @@ type WorkoutLog struct {
 	UserID       uuid.UUID
 	Exercises    []ExerciseLog
 	Timestamp    int64
-	SleepHours   int `json:"sleepHours"` // Hours slept before workout
+	SleepHours   int `json:"sleepHours"`   // Hours slept before workout
 	SleepQuality int `json:"sleepQuality"` // 1-10 scale
-	StressLevel  int `json:"stressLevel"` // 1-10 scale
+	StressLevel  int `json:"stressLevel"`  // 1-10 scale
 }
 
 // RecoveryMetrics tracked daily
@@ -37,5 +37,3 @@ type RecoveryMetrics struct {
 	StressLevel  int // 1-10
 	Soreness     int // 1-10 DOMS rating
 }
-
-
